Skip the actual commit of the merge
The `--no-commit` fixes the command exiting because of merge requiring
an editor to commit the results:

```
error: Terminal is dumb, but EDITOR unset
Not committing merge; use 'git commit' to complete the merge.
```

The `--no-ff` makes sure the merge is always treated as a merge and
not as a "fast-forward". This ensures consistency of the operation.