Handle detached head checkout for non-local parent branch
When checking out a non-local $PARENT_BRANCH (i.e. origin/master, PR
branches, etc.) the working directory is left in a 'detached HEAD'
state. This can cause potential issues in the following commands.

To assure the checkout state is clean, a local branch is created when
checking out the $PARENT_BRANCH.