[semver:minor] update merge with parent (#28)
* Set git identity to a sane default, if not already set

`git merge` command requires the user identity to be set, otherwise
the execution of `merge-with-parent` breaks with an error:

```
*** Please tell me who you are.

Run

  git config --global user.email "you@example.com"
  git config --global user.name "Your Name"

to set your account's default identity.
Omit --global to set the identity only in this repository.

fatal: unable to auto-detect email address (got 'root@4f4eea8052cd.(none)')
```

* Skip the actual commit of the merge

The `--no-commit` fixes the command exiting because of merge requiring
an editor to commit the results:

```
error: Terminal is dumb, but EDITOR unset
Not committing merge; use 'git commit' to complete the merge.
```

The `--no-ff` makes sure the merge is always treated as a merge and
not as a "fast-forward". This ensures consistency of the operation.

* Handle detached head checkout for non-local parent branch

When checking out a non-local $PARENT_BRANCH (i.e. origin/master, PR
branches, etc.) the working directory is left in a 'detached HEAD'
state. This can cause potential issues in the following commands.

To assure the checkout state is clean, a local branch is created when
checking out the $PARENT_BRANCH.

* Parameterize optional checkout and merge flags